make inflight parsing throw an error message when numeric value specified for a subscriber (where it is inappropriate.) #316

+++ sarra/sr_util.py
@@ -470,27 +470,30 @@ class sr_transport():
                 if parent.inflight == None or msg.partflg == 'i' :
                    self.get(remote_file,new_file,remote_offset,msg.local_offset,msg.length)
 
-                elif parent.inflight == '.' :
-                   new_lock = '.' + new_file
-                   self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
-                   if os.path.isfile(new_file) : os.remove(new_file)
-                   os.rename(new_lock, new_file)
-                      
-                elif parent.inflight[0] == '.' :
-                   new_lock  = new_file + parent.inflight
-                   self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
-                   if os.path.isfile(new_file) : os.remove(new_file)
-                   os.rename(new_lock, new_file)
+                elif type(parent.inflight) == str :
+                   if parent.inflight == '.' :
+                       new_lock = '.' + new_file
+                       self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
+                       if os.path.isfile(new_file) : os.remove(new_file)
+                       os.rename(new_lock, new_file)
+                    
+                   elif parent.inflight[0] == '.' :
+                       new_lock  = new_file + parent.inflight
+                       self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
+                       if os.path.isfile(new_file) : os.remove(new_file)
+                       os.rename(new_lock, new_file)
 
-                elif parent.inflight[-1] == '/' :
-                   try :  
-                          os.mkdir(parent.inflight)
-                          os.chmod(parent.inflight,parent.chmod_dir)
-                   except:pass
-                   new_lock  = parent.inflight + new_file
-                   self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
-                   if os.path.isfile(new_file) : os.remove(new_file)
-                   os.rename(new_lock, new_file)
+                   elif parent.inflight[-1] == '/' :
+                       try :  
+                              os.mkdir(parent.inflight)
+                              os.chmod(parent.inflight,parent.chmod_dir)
+                       except:pass
+                       new_lock  = parent.inflight + new_file
+                       self.get(remote_file,new_lock,remote_offset,msg.local_offset,msg.length)
+                       if os.path.isfile(new_file) : os.remove(new_file)
+                       os.rename(new_lock, new_file)
+                else:
+                    self.logger.error('inflight setting: %s, not for remote.' % parent.inflight )
 
                 self.logger.debug('proto.checksum={}, msg.sumstr={}'.format(proto.checksum, msg.sumstr))
                 msg.onfly_checksum = proto.get_sumstr()
